services/ticket-log-service/requirements.txt
flask==3.1.3
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.5
